Remove outdated kill method to stop sidecar
Sending the kill signal was needed for Istio < v1.3 and 1.3 has been
released 4 years ago.

If the `ENVOY_ADMIN_API` is set to the default port `15000` the
`ISTIO_QUIT_API` configured correctly by default.
On deployments that restart the containers the shutdown behaviour might
be undesirable.
The Istio sidecar will be restarted but when the pod has multiple
containers that means that meaningful work in the containers that did
not crash could be interrupted.

So for pods that have multiple containers that need traffic through the
service mesh, it is recommended to set `NEVER_KILL_ISTIO` to `true`.